shield_product/requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==2.1.1
pillow==10.4.0
fastapi==0.115.0
uvicorn==0.30.6
celery[redis]==5.4.0
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
